Divided shell in two
